+++ test/test.js
@@ -23,7 +23,7 @@ describe('etag(entity)', function () {
 
     it('should work containing Unicode', function () {
       assert.equal(etag('论'), '"3-aW9HeLTk2Yt6lf7zJYElgw"')
-      assert.equal(etag('论', {weak: true}), 'W/"3-438093ff"')
+      assert.equal(etag('论', {weak: true}), 'W/"3-aW9HeLTk2Yt6lf7zJYElgw"')
     })
 
     it('should work for empty string', function () {
@@ -84,19 +84,23 @@ describe('etag(entity)', function () {
 
     describe('when "true"', function () {
       it('should generate a weak ETag for a string', function () {
-        assert.equal(etag('', {weak: true}), 'W/"0-0"')
-        assert.equal(etag('beep boop', {weak: true}), 'W/"9-7f3ee715"')
+        assert.equal(etag('', {weak: true}), 'W/"0-1B2M2Y8AsgTpgAmY7PhCfg"')
+        assert.equal(etag('beep boop', {weak: true}), 'W/"9-Z34SGyQ2IB7YzB7HMkCjrQ"')
         assert.equal(etag(str5kb, {weak: true}), 'W/"1400-8Kq68cJq4i+5US7RLWrE1g"')
       })
 
       it('should generate a weak ETag for a Buffer', function () {
-        assert.equal(etag(new Buffer(0), {weak: true}), 'W/"0-0"')
-        assert.equal(etag(new Buffer([1, 2, 3]), {weak: true}), 'W/"3-55bc801d"')
+        assert.equal(etag(new Buffer(0), {weak: true}), 'W/"0-1B2M2Y8AsgTpgAmY7PhCfg"')
+        assert.equal(etag(new Buffer([1, 2, 3]), {weak: true}), 'W/"3-Uonfc331cyb83SJZevsfrA"')
         assert.equal(etag(buf5kb, {weak: true}), 'W/"1400-8Kq68cJq4i+5US7RLWrE1g"')
       })
 
       it('should generate a weak ETag for fs.Stats', function () {
         assert.ok(isweak(etag(fs.statSync(__filename), {weak: true})))
+      })
+
+      it('should generate different ETags for different strings', function () {
+        assert.notEqual(etag('plumless', {weak: true}), etag('buckeroo', {weak: true}))
       })
     })
   })