+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Use MD5 instead of MD4 in weak ETags over 1KB
+
 1.6.0 / 2015-05-10
 ==================