+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Improve stat performance by removing hashing
+  * Remove base64 padding in ETags to shorten
   * Use MD5 instead of MD4 in weak ETags over 1KB
 
 1.6.0 / 2015-05-10