+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Always include entity length in ETags for hash length extensions
+  * Generate non-Stats ETags using MD5 only (no longer CRC32)
   * Improve stat performance by removing hashing
   * Remove base64 padding in ETags to shorten
   * Use MD5 instead of MD4 in weak ETags over 1KB